+++ locale/pt-br/setting.lua
@@ -97,6 +97,8 @@ config.diagnostics.ignoredFiles.Disable  = -- TODO: need translate!
 "These files are not diagnosed."
 config.diagnostics.disableScheme  = -- TODO: need translate!
 'Do not diagnose Lua files that use the following scheme.'
+config.diagnostics.unusedLocalExclude = -- TODO: need translate!
+'Do not diagnose `unused-local` when the variable name matches the following pattern.'
 config.workspace.ignoreDir        = -- TODO: need translate!
 "Ignored files and directories (Use `.gitignore` grammar)."-- .. example.ignoreDir,
 config.workspace.ignoreSubmodules = -- TODO: need translate!